requirements.txt
django
djangorestframework
environs
drf-yasg
psycopg2
